requirements.txt
flask
flask-cors
openai>=1.13.3
python-dotenv
gunicorn
